requirements.txt
# 核心相依套件
feedparser>=6.0.10
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
openai>=1.12.0
loguru>=0.7.2

# 開發/測試（可選）
pytest>=7.4.0
